requirements.txt
-i https://pypi.org/simple
blinker==1.8.2; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
flask==3.0.3; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
werkzeug==3.0.3; python_version >= '3.8'
